+++ sistema_gestao/database/migrations/2025_02_05_183854_create_colaboradors_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->increments('id');
             $table->string('nome'); 
             $table->string('email');
-            $table->integer('cpf');
+            $table->bigInteger('cpf');
             $table->integer('unidade_id')->unsigned()->nullable();
             $table->timestamps();
         });
